pyproject.toml
[project]
name = "llm-agent-template"
version = "0.1.0"
description = "LLM Agent Template"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "python-multipart",  # for file uploads
    "uvicorn[standard]",
    "python-dotenv",
    "openai>=1.0.0",
    "pypdf",  # for PDF processing
    "PyPDF2>=3.0.0",
    "numpy>=1.24.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.22",
    "langchain-openai>=0.0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
] 

[tool.hatch.build.targets.wheel]
packages = ["backend"]
